feat: add abstract controller and finish todo
