+++ composer.json
@@ -15,7 +15,8 @@
     "symfony/filesystem": "^3.0",
     "doctrine/collections": "~1.3.0",
     "segmentio/analytics-php": "^1.8",
-    "symfony/dotenv": "^4.4"
+    "symfony/dotenv": "^4.4",
+    "symfony/http-foundation": "^4.4"
   },
   "require-dev": {
     "phpunit/phpunit": "^5",
