+++ controllers/admin/self-managed/WelcomeController.php
@@ -2,42 +2,20 @@
 
 namespace PrestaShop\Module\AutoUpgrade\Controller;
 
-use PrestaShop\Module\AutoUpgrade\UpgradeContainer;
 use PrestaShop\Module\AutoUpgrade\BackupFinder;
+use PrestaShop\Module\AutoUpgrade\Controller\AbstractPageController;
 
-class WelcomeController /* Future abstract class */
+class WelcomeController extends AbstractPageController
 {
-    // TODO: Move this in a parent class
-
-    /** @var UpgradeContainer */
-    protected $upgradeContainer;
-
-    public function __construct(UpgradeContainer $upgradeContainer)
-    {
-        // TODO: add twig into class variable
-        $this->upgradeContainer = $upgradeContainer;
-    }
-    // END OF TODO
-
     public function index(): string
     {
-        $psVersion = $this->upgradeContainer->getProperty($this->upgradeContainer::PS_VERSION);
-        $psClass = '';
         $backupPath = $this->upgradeContainer->getProperty($this->upgradeContainer::BACKUP_PATH);
         $backupFinder = new BackupFinder($backupPath);
 
-        if (version_compare($psVersion, '1.7.8.0', '<')) {
-            $psClass = 'v1-7-3-0';
-        } else if (version_compare($psVersion, '9.0.0', '<')){
-            $psClass = 'v1-7-8-0';
-        }
-
-        return $this->upgradeContainer->getTwig()->render(
-            '@ModuleAutoUpgrade/layouts/layout.html.twig',
+        return $this->renderPage(
+            'welcome',
             [
-                'page' => 'welcome',
-                'ps_version' => $psClass,
-                'empty_backup' => empty($backupFinder->getAvailableBackups()),
+                'empty_backup' => empty($backupFinder->getAvailableBackups())
             ]
         );
     }
